ChatBot/PythonAnywhere/requirements.txt
twilio==6.42.0
PyGithub==1.51
pandas==1.0.4
numpy==1.18.4
